Add initial implementation of websocket authentication with unit tests
This patch makes a number of modifications to appSession in order to
avoid duplication in the implementation of attachSessionToHttpRequestIfExists,
a new function which is now also returned from the main appSession function.

It also modifies the auth middleware to conditionally return both the
router and an authentication function which can be used as in the ws
package code example here https://github.com/websockets/ws#client-authentication.