Issue 227 - allow custom session id generation
This commit address [issue
227](https://github.com/auth0/express-openid-connect/issues/227)
allowing an implementation of genid to be used when a custom session
store is used.  This is fully compatibile with the existing
implementation and aligns with how expression-session allows for
overriding the way session ids are generated. Note because of how
session ids are generated in this library, it is not possible to simply
fall into the custom store's genid function.

In implementing these changes, the following was done:
- Moved the default session id generation into config to make it clear
what said default is
- Added a new parameter genid compatible with how it works for
express-session
- Updated the appsession code to take advantage of the provided value and
fallback to the default.
- Added unit tests validating the genid parameter being used when
specified and not having any effect on the encrypted cookie
implementation.
- Reordered the spread operators in the appsession custom store tests
so that overrides are respected.  This changed no other existing tests.
- Updated documentation with the additional, optional parameter along
with links to the compatabile usage in express-session.